Update workflow actions to latest
Fix various deprecation warnings.  The actions-rs and actions/*release*
workflows are unmaintained, so replace them with substitutes.

Use generic content type for uploaded artifacts.  This avoids the need to
run the release step once per artifact, and the content types don't seem
to be attached to the final download anyway.

Automatically generate release notes, but leave the release in draft
so the release notes can be cleaned up before releasing.