Update workflow actions to latest
Fix various deprecation warnings.  The actions-rs and actions/*release*
workflows are unmaintained; replace the former with manual Rust commands
and the latter with a substitute.

Use generic content type for uploaded artifacts.  This avoids the need to
run the release step once per artifact, and the content types don't seem
to be attached to the final download anyway.